% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_db_columns.R
\name{check_db_columns}
\alias{check_db_columns}
\title{Check Column Meta Data}
\usage{
check_db_columns(ref)
}
\arguments{
\item{ref}{\link[loadcheckr]{check_db_references} object that holds the connections and tables that you are comparing}
}
\value{
dataframe listing out the results of the test
}
\description{
Checks the column names, column lengths and data types in both data tables
}
