% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_db_keys.R
\name{check_db_keys}
\alias{check_db_keys}
\title{Check Keys}
\usage{
check_db_keys(ref)
}
\arguments{
\item{ref}{\link[loadcheckr]{check_db_references} object that holds the connections and tables that you are comparing}
}
\value{
dataframe listing out the results of the test
}
\description{
Checks whether the same records appear in both data tables based on a set of keys
}
