% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_db_random_sample.R
\name{check_db_random_sample}
\alias{check_db_random_sample}
\title{Check Random Row Sample}
\usage{
check_db_random_sample(conn_a, table_name_a, conn_b, table_name_b, key_field,
  key_field_index, num_rows)
}
\arguments{
\item{conn_a}{the first dbr data connection}

\item{conn_b}{the second dbr data connection}

\item{key_field}{what you are using as a primary key to compare rows, if you make a composite put it in parenthesis (person + date)}

\item{num_rows}{the number of rows that you want to look at}

\item{table_a}{the first table name}

\item{table_b}{the second table name}
}
\value{
output of a datapointsr::show_values object
}
\description{
Checks all the content in a randomly selected set of rows. This function ignores data column case.
}

