% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_db_references.R
\name{check_db_references}
\alias{check_db_references}
\title{Holds References to Connections}
\usage{
check_db_references(conn_x, schema_x, table_name_x, conn_y, schema_y,
  table_name_y, keys = NULL)
}
\arguments{
\item{conn_x}{the first dbr data connection}

\item{schema_x}{the first database schema name}

\item{conn_y}{the second dbr data connection}

\item{schema_y}{the second database schema name}

\item{table_x}{the first table name}

\item{table_y}{the second table name}
}
\value{
a list of dbr data connections and table names that can be compared
}
\description{
Holds references to both data connections and the tables being compared
}
