% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_db_row_keys_match.R
\name{check_db_row_keys_match}
\alias{check_db_row_keys_match}
\title{Check All Records}
\usage{
check_db_row_keys_match(conn_a, table_name_a, conn_b, table_name_b, key_field)
}
\arguments{
\item{conn_a}{the first dbr data connection}

\item{conn_b}{the second dbr data connection}

\item{key_field}{what you are using as a primary key to compare rows, if you make a composite put it in parenthesis (person + date)}

\item{table_a}{the first table name}

\item{table_b}{the second table name}
}
\value{
output of a datapointsr::show_values object
}
\description{
Checks whether the same records appear in both data tables based on a primary key
}

