% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_db_rows.R
\name{check_db_rows}
\alias{check_db_rows}
\title{Check Row Counts}
\usage{
check_db_rows(ref)
}
\arguments{
\item{ref}{\link[loadcheckr]{check_db_references} object that holds the connections and tables that you are comparing}
}
\value{
dataframe listing out the results of the test
}
\description{
Checks the number of rows in both data tables
}

