% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_db_columns.R
\name{check_db_columns}
\alias{check_db_columns}
\title{Check Columns}
\usage{
check_db_columns(conn_a, table_name_a, conn_b, table_name_b,
  ignore_case = FALSE)
}
\arguments{
\item{conn_a}{the first dbr data connection}

\item{conn_b}{the second dbr data connection}

\item{ignore_case}{ignore case when comparing column names}

\item{table_a}{the first table name}

\item{table_b}{the second table name}
}
\value{
a list of objects describing the results, prints out to the log as well
}
\description{
Checks whether the data sources agree on data types: any numeric == any numeric, any character == any character
}
\details{
Checks the column names and data types in both data tables
}

