% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_db_sample.R
\name{check_db_random_sample}
\alias{check_db_random_sample}
\title{Check Random Row Sample}
\usage{
check_db_random_sample(ref, num_rows)
}
\arguments{
\item{ref}{\link[loadcheckr]{check_db_references} object that holds the connections and tables that you are comparing}

\item{num_rows}{the number of rows that you will randomly sample}
}
\value{
dataframe listing out the results of the test
}
\description{
Checks all the content in a randomly selected set of rows. This function ignores data column case.
}
