% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_db_rows.R
\name{check_db_rows}
\alias{check_db_rows}
\title{Check Row Counts}
\usage{
check_db_rows(conn_a, table_name_a, conn_b, table_name_b)
}
\arguments{
\item{conn_a}{the first dbr data connection}

\item{conn_b}{the second dbr data connection}

\item{table_a}{the first table name}

\item{table_b}{the second table name}
}
\value{
a list of objects describing the results, prints out to the log as well
}
\description{
Checks the number of rows in both data tables
}

